CHR-22 Move mappers from endpoints to separate mappers.go file
